Move mcu bringup tasks to shared lib file; update Makefile to init data to DCCM and build libs
